requirements-dev.txt
pytest
pytest-django
pytest-cov
pytest-xdist
flake8
responses
djangorestframework
django-filter
